tests: add separate pmemcheck test for append_reserve
data integrity check should not be extended with multiple functional checks.

This test found a bug, so it's delivered already with a fix:
use persist_size in span_create_entry, instead of improper value.
